+++ dolomite_engine/hf_models/models/gpt_ensemble/normalization.py
@@ -31,3 +31,24 @@ class EnsembleRMSNorm(nn.RMSNorm):
         input = weight * input
 
         return input
+
+
+_NORMALIZATION_FUNCTIONS = {"rmsnorm": EnsembleRMSNorm}
+
+
+def get_ensemble_normalization_function(
+    name: str,
+    normalized_shape: int,
+    tp_world_size: int,
+    eps: float = 1e-5,
+    normalization_implementation: str = "torch",
+) -> nn.LayerNorm:
+    if name in _NORMALIZATION_FUNCTIONS:
+        return _NORMALIZATION_FUNCTIONS[name](
+            normalized_shape,
+            tp_world_size=tp_world_size,
+            eps=eps,
+            normalization_implementation=normalization_implementation,
+        )
+
+    raise ValueError(f"unexpected `normalization_function` {name}")
